requirements.txt
# Core ML/NLP Libraries
scikit-learn==1.3.2
numpy==1.24.3
scipy==1.11.4

# NLP & Embeddings
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1

# Text Processing
regex==2023.10.3

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Environment & Configuration
python-dotenv==1.0.0

# Utilities
tqdm==4.66.1
joblib==1.3.2
